backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
celery>=5.4
redis>=5.0
sqlalchemy>=2.0
alembic>=1.13
pydantic>=2.7
pydantic-settings>=2.3
email-validator>=2.1
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.9
cryptography>=42
openpyxl>=3.1
lxml>=5.2
anthropic>=0.28
deepl>=1.18
google-cloud-translate>=3.15
requests>=2.32
python-dotenv>=1.0
aiofiles>=23.2
sse-starlette>=2.1
